[UPDATE] 댓글 백엔드 api 구현, 이외 일부 수정

+++ noriteo-back/noriteo/src/main/java/com/ptu/noriteo/model/BoardPic.java
@@ -7,9 +7,9 @@ public class BoardPic {
     private Long boardPicId;
     private Long boardId;
     private String boardPicUrl;
-    private Integer boardPicOrder;
+    private Long boardPicOrder;
 
-    public BoardPic(Long boardId, String boardPicUrl, int boardPicOrder) {
+    public BoardPic(Long boardId, String boardPicUrl, Long boardPicOrder) {
         this.boardId = boardId;
         this.boardPicUrl = boardPicUrl;
         this.boardPicOrder = boardPicOrder;
